+++ CHANGELOG.md
@@ -3,10 +3,11 @@ Change log for Simpl Cloud
 
 A list of notable changes to the Simpl-Cloud library included in each release.
 
-1.1.1 (unreleased)
-==================
+1.2 (unreleased)
+================
 
-- Nothing changed yet.
+- Implement logic for restricting teams to a minimum/maximum player count
+  (including balancing).
 
 
 1.1.0.1 (13 June 2022)